refactor:next/image사용, api요청 분리, setIsOpen 수정
