--- lib/axios.ts
@@ -1,7 +1,0 @@
-import axios from "axios";
-
-const instance = axios.create({
-  baseURL: "https://panda-market-api.vercel.app",
-});
-
-export default instance;
